requirements.txt
altair==5.5.0
anyio==4.10.0
appnope==0.1.4
arrow==1.3.0
asttokens==3.0.0
attrs==25.3.0
binaryornot==0.4.4
blinker==1.9.0
branca==0.8.1
cachetools==6.1.0
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
comm==0.2.3
contourpy==1.3.0
cookiecutter==2.6.0
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.59.0
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
humanize==4.12.3
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==8.18.1
ipyvue==1.11.2
ipyvuetify==1.11.3
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
Mesa==2.4.0
Mesa-Viz-Tornado==0.1.3
narwhals==2.0.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.0.2
osmnx==2.0.5
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pydeck==0.9.1
Pygments==2.19.2
pymdown-extensions==10.16.1
pyogrio==0.11.1
pyparsing==3.2.3
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
reacton==1.9.1
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rich-click==1.8.9
rpds-py==0.26.0
shapely==2.0.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
solara==1.50.1
solara-server==1.50.1
solara-ui==1.50.1
stack-data==0.6.3
starlette==0.47.2
streamlit==1.47.1
tenacity==9.1.2
text-unidecode==1.3
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250708
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
widgetsnbextension==4.0.14
xyzservices==2025.4.0
zipp==3.23.0
scikit-learn==1.5.0
folium==0.14.0
streamlit-folium==0.16.0
